Refactored DateLabel common stuff
