+++ TruckFeed/TruckFeed/EndDateCell.swift
@@ -13,44 +13,22 @@ class EndDateSwitchCell: UITableViewCell {
 
     @IBOutlet weak var endDateLabel:UILabel!
     public var delegate: AddScheduleDataProvider?
+    private var dateLabel: DateLabel?
     lazy var realm = try? Realm()
     
     override func awakeFromNib() {
-        self.updateDateLabel()
+        self.dateLabel = DateLabel()
+        let schedule = self.dateLabel?.getDateFromSchedule()
+        endDateLabel.text = schedule?.endDate
+    
         NotificationCenter.default.addObserver(forName: NSNotification.Name(rawValue: "com.lovelejess.endDateLabelSelected"), object: nil, queue: nil, using: updateEndDateLabel)
     }
     
     private func updateEndDateLabel(notification: Notification) -> Void {
         if let userInfo = notification.userInfo {
             if let date = userInfo["date"]  as? String {
-                self.saveDateLabel(date: date)
+                self.dateLabel?.saveDateLabel(date: date)
             }
-        }
-    }
-    
-    func updateDateLabel() {
-        if let scheduleRealm = realm?.objects(Schedule.self) {
-            let schedule = scheduleRealm.first
-            endDateLabel.text = schedule?.endDate
-        }
-    }
-    
-    func saveDateLabel(date:String) {
-        if let schedule = realm?.objects(Schedule.self).first {
-            try! self.realm?.write {
-                schedule.endDate = date
-            }
-        }
-        else {
-            self.addSchedule(date: date)
-        }
-    }
-    
-    private func addSchedule(date: String) {
-        let schedule = Schedule()
-        schedule.endDate = DateUtility.getCurrentDateTime()
-        try! self.realm?.write {
-            realm?.add(schedule)
         }
     }
 }
